Add drain hand as a recognized gesture
The rain hand included in version 2.2 is a nice interactive addition.
But, how to remove all that water precipitated into the model?

Introducing: the drain hand.  A hand, palm down, with two fingers extended
and the other three fingers curled under and out of view of the depth
camera.

Think of this as representing a map contour line where it crosses a river;
the water flows out of the V.  The two extended finger tips must have a
separation distance around the length of the fingers.  Imagine a virtual
equilateral triangle with two sides defined by the hand's two extended
fingers.

Internally, this adds a direction member to the Hand struct.  The default
direction is for rain to fall.  When a drain hand is detected, the rain
evaporates.  Specifically, a drain hand causes rain to "un-fall" at
double the rainStrength value.

It is possible to have water fall as rain at a higher land elevation,
flow downhill, and drain from some catchment area molded in the sand.

Signed-off-by: Sean Robinson <sean.robinson@scottsdalecc.edu>