Add a pause before exit to allow error messages to be read
This script runs in a shell window that disappears when the script ends.
Without waiting for the Enter key, no error feedback gets to the user.

This bit me when there were file permission errors I was not aware of
because the error message disappeared before I could read it.

Signed-off-by: Sean Robinson <sean.robinson@scottsdalecc.edu>