Add rainStrength control pipe command
This replicates the -rs command line argument as a control pipe command
so that the local rainfall rate can be changed in a running sandbox.

Because negative rain strength values are accepted, another way to use this
control pipe command is for local drain/evaporation.  That is, with a
positive rainStrength, an open hand gesture causes "water" to form in the
sandbox model.  But, with a negative rainStrength, an open hand gesture
removes water locally.

All "open hands" recognized by the sandbox are treated in the same
manner.  So, this addition does not allow some hands to rain and others
to evaporate, simultaneously.  With a way to toggle positive and negative
rainStrength values, spread-finger hands can be toggled between rain and
evaporation use.

Signed-off-by: Sean Robinson <sean.robinson@scottsdalecc.edu>